+++ test/test_chunkstore.py
@@ -185,7 +185,7 @@ class SentinelHubStore3DTestWithNonDefaultCrs(SentinelHubStoreTest):
     def get_cube_config(self):
         crs_name = 'EPSG:3035'
         crs = pyproj.CRS.from_string(crs_name)
-        t = pyproj.Transformer.from_crs('CRS84', crs)
+        t = pyproj.Transformer.from_crs('OGC:CRS84', crs)
         (x1, y1), (x2, y2) = t.transform((10.2, 10.3), (53.5, 53.6))
         spatial_res = (x2 - x1) / 4000
         y2 = y1 + spatial_res * 4000